Reorganized project files.
